nodes/storage: export filesystems only once
Upgrading the zfs service was dangerous primarily on nasbox, since the
service always reshared all filesystems using zfs share, which meant
downtime for nfs clients until the command finished. Now the filesystems
will be shared only once on boot. Future restarts of the service will
not cause the filesystems to be reshared.