clusterconf/monitor: remove indecisive rules for smartctl
Rules for raw read error rate and reallocated sectors fire often, yet do
not signify an immediate problem. They're thus lost in the churn and
mostly ignored. We will instead rely on rules for zpool status and error
counters on individual vdevs. At that time, it's clear that the disk
needs to be replaced.