int.log: disable mbuffer
mbuffer passes data in blocks. The larger the buffer, the larger block
size. mbuffer will pass data only when at least one block is full, so
when it is several MB in size, it delays log message processing. Block
size is configurable, but a warning is issued if there's too many of
them... disabling until a solution is found.