configs/dns-resolver: remove NAT IPs from resolvers in brq
The monitoring has a private IPv4 address in Prague, so access to DNS
resolvers in Brno required us to allow connections from NAT in Prague.
The NAT now has new addresses which are from subnets registered
in vpsAdmin, so it is automatically allowed and we don't need to do
anything more.