Reintroduce production channel
a1605779a09553dd34e7f2fffe84aaae7889da13 removed channel prod-24.07 and
we tried to use only staging with configurable stable/unstable package
sets. That turned out to be insufficient for osctl, as it is a part of
vpsAdminOS repository and we cannot switch between versions like we do
with kernel and ZFS; there is always only one version of osctl.

For now, the production channel is pointing to the staging git branch
and we expect it to be on par with the staging channel. The difference
will be that now we can arbitratily switch the staging channel to
development versions without worrying about production nodes.