nodes: add netboot-kexec script to initrd
This will allow us to easily kexec from initrd if the current one will
fail to boot the system for some reason. We'll be able to deploy a fix
on the netboot server and then kexec the node into it.

It's a simplified variant of netboot-kexec from confctl, it fetches the
current generation from the netboot server by default.