monitor: replace QuietHours rule with alertmanager time intervals
When we started using Prometheus/alertmanager, it didn't have a feature
to mute alerts based on the time, so we used a custom Prometheus rule
QuietHours that fired during the night and it was used to inhibit critical
alerts within alertmanager.

This patch replaces QuietHours with time intervals configured in
alertmanager. This solution allows us to choose different work/quiet
hours for each admin.